- fix for ATMEL AVR GCC which does not seem to support the C++ standard library headers (https://stackoverflow.com/questions/16276293/c-standard-compliance-in-avr-gcc)
